refactor(server): adicionando as funções do app para /api/app.js
